modify the dependencies * delete class_macros.rb and loader.rb * add 'ActiveSupport' dependency in gemspec * modify base.rb and core.rb
